one hot encode

+++ supervised_learning/0x01-classification/24-one_hot_encode.py
@@ -19,7 +19,7 @@ def one_hot_encode(Y, classes):
     mx = np.zeros((classes, m))
     i = 0
     for n in Y:
-        if type(n) is not int:
+        if type(n) is not np.int64:
             return None
         mx[n:n+1, i:i+1] = 1
         i += 1
